Rename driver files and make key standardization a method
